categories front end guards and policies on be

+++ app/Http/Controllers/Api/Dashboard/CategoryController.php
@@ -15,6 +15,7 @@ use App\Interfaces\ImageServiceInterface;
 use App\Http\Resources\CategoryCollection;
 use App\Http\Requests\CategoryStoreRequest;
 use App\Http\Requests\CategoryUpdateRequest;
+use Illuminate\Auth\Access\AuthorizationException;
 use App\Http\Resources\Category as CategoryResrouce;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
@@ -51,7 +52,7 @@ class CategoryController extends Controller
 	public function store(CategoryStoreRequest $request)
 	{
 		try {
-			$request->user()->can('create', Category::class);
+			$this->authorize('create', Category::class);
 
 			$input = $request->validated();
 
@@ -95,15 +96,18 @@ class CategoryController extends Controller
 			Cache::forget('categories');
 
 			return response()->json($response, 201);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
 		} catch (QueryException $qex) {
 			dd($qex);
 			$errorCode = $qex->errorInfo[1];
-			if ($errorCode  == 1062) {
-				return response()->json(['error' => 'Category already exists'], 500);
+			if ($errorCode == 1062) {
+				return response()->json(['message' => 'Category already exists'], 500);
 			}
+			return response()->json(['message' => 'Something went wrong, try again later'], 500);
 		} catch (\Exception $ex) {
 			dd($ex);
-			return response()->json(['error' => 'Something went wrong, try again later'], 500);
+			return response()->json(['message' => 'Something went wrong, try again later'], 500);
 		}
 	}
 
@@ -116,36 +120,48 @@ class CategoryController extends Controller
 	 */
 	public function update(CategoryUpdateRequest $request, $id)
 	{
-		$request->user()->can('update', Category::class);
 
-		$category = Category::withTrashed()->findOrFail($id);
+		try {
+			$this->authorize('update', Category::class);
 
-		$input = $request->validated();
+			$category = Category::withTrashed()->findOrFail($id);
 
-		$responseData = null;
+			$input = $request->validated();
 
-		if ($request->has('name')) {
-			$category->slug = null;
+			$responseData = null;
+
+			if ($request->has('name')) {
+				$category->slug = null;
+			}
+
+			if ($request->has('parentId')) {
+				$input['parent_id'] = $request->parentId;
+				$parentCategory = Category::withTrashed()->findOrFail($request->parentId);
+				$responseData['parentName'] = $parentCategory->name;
+				$lastPosition =  $this->getLastPosition($request->parentId);
+			} else {
+				$lastPosition =  $this->getLastPosition();
+			}
+
+			$input['position'] =  $lastPosition + 1;
+			$category->update($input);
+
+			$responseData['message'] = "Category updated";
+			$responseData['position'] = $input['position'];
+
+			Cache::forget('categories');
+
+			return response()->json($responseData, 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
+		} catch (ModelNotFoundException $e) {
+			DB::rollBack();
+			debug($e);
+			return response()->json(['message' => 'Category not found'], 500);
+		} catch (\Exception $e) {
+			DB::rollBack();
+			return response()->json(['message' => 'Something went wrong, try again later'], 500);
 		}
-
-		if ($request->has('parentId')) {
-			$input['parent_id'] = $request->parentId;
-			$parentCategory = Category::withTrashed()->findOrFail($request->parentId);
-			$responseData['parentName'] = $parentCategory->name;
-			$lastPosition =  $this->getLastPosition($request->parentId);
-		} else {
-			$lastPosition =  $this->getLastPosition();
-		}
-
-		$input['position'] =  $lastPosition + 1;
-		$category->update($input);
-
-		$responseData['message'] = "Category updated";
-		$responseData['position'] = $input['position'];
-
-		Cache::forget('categories');
-
-		return response()->json($responseData, 200);
 	}
 
 	/**
@@ -154,11 +170,12 @@ class CategoryController extends Controller
 	 * @param  int  $id
 	 * @return \Illuminate\Http\Response
 	 */
-	public function destroy(Request $request, $id)
+	public function destroy($id)
 	{
-		$request->user()->can('forceDelete', Category::class);
 
 		try {
+			$this->authorize('forceDelete', Category::class);
+
 			DB::beginTransaction();
 
 			$category = Category::withTrashed()->with('subCategories')->findOrFail($id);
@@ -187,6 +204,10 @@ class CategoryController extends Controller
 			DB::commit();
 
 			return response()->json(['message' => 'Category ' . $category->name . ' was deleted'], 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
+		} catch (ModelNotFoundException $e) {
+			return response()->json(['message' => 'Category not found'], 404);
 		} catch (\Illuminate\Database\QueryException $e) {
 			DB::rollBack();
 			debug($e);
@@ -197,34 +218,47 @@ class CategoryController extends Controller
 		}
 	}
 
-	public function disable(Request $request, $id)
+	public function disable($id)
 	{
+		try {
+			$this->authorize('disable', Category::class);
 
-		$request->user()->can('disable', Category::class);
+			$category = Category::withTrashed()->findOrFail($id);
 
-		$category = Category::withTrashed()->findOrFail($id);
-
-		$category->delete();
-		return response()->json(['message' => 'Category ' . $category->name . ' was disabled', 'deletedAt' => $category->deleted_at], 200);
+			$category->delete();
+			return response()->json(['message' => 'Category ' . $category->name . ' was disabled', 'deletedAt' => $category->deleted_at], 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
+		} catch (ModelNotFoundException $e) {
+			return response()->json(['message' => 'Category not found'], 404);
+		} catch (Exception $e) {
+			return  response()->json(['message' => $e->getMessage()], 500);
+		}
 	}
 
-	public function restore(Request $request, $id)
+	public function restore($id)
 	{
-		$request->user()->can('restore', Category::class);
+		try {
+			$this->authorize('restore', Category::class);
 
-		$category = Category::withTrashed()->findOrFail($id);
+			$category = Category::withTrashed()->findOrFail($id);
 
-		$category->restore();
+			$category->restore();
 
+			Cache::forget('categories');
 
-		Cache::forget('categories');
-
-		return response()->json(['message' => 'Category ' . $category->name . ' was restored'], 200);
+			return response()->json(['message' => 'Category ' . $category->name . ' was restored'], 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
+		} catch (ModelNotFoundException $e) {
+			return response()->json(['message' => 'Category not found'], 404);
+		} catch (Exception $e) {
+			return  response()->json(['message' => $e->getMessage()], 500);
+		}
 	}
 
 	public function search($catagoryName)
 	{
-
 		$categories = Category::with(['discount', 'products', 'subProducts', 'subCategories' => function ($query) {
 			$query->withTrashed();
 			$query->orderBy('position', 'asc');
@@ -241,6 +275,8 @@ class CategoryController extends Controller
 	public function removeParent($id)
 	{
 		try {
+			$this->authorize('update', Category::class);
+
 			$category = Category::withTrashed()->findOrFail($id);
 
 			$category->parent_id = null;
@@ -250,8 +286,9 @@ class CategoryController extends Controller
 			$category->save();
 
 			return response()->json(['message' => 'Parent category removed', 'position' => $category->position], 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
 		} catch (ModelNotFoundException $mnfe) {
-			debug($mnfe);
 			return response()->json(['message' => 'Category not found'], 404);
 		} catch (\Exception $e) {
 			debug($e);
