+++ app/Http/Controllers/Api/Dashboard/CategoryImageController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Storage;
 use App\Interfaces\ImageServiceInterface;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class CategoryImageController extends Controller
 {
@@ -22,14 +23,18 @@ class CategoryImageController extends Controller
 	public function store(Request $request, $id)
 	{
 		try {
+
+			$this->authorize('update', Category::class);
+
 			$storagePath = '/categories_images/' . $id;
 
 			$path = $this->imageService->storeImage($request->file('image'), $storagePath, 'categories', 'image', intval($id));
 
 			return response()->json(['imagePath' => $path], 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
 		} catch (Exception $e) {
 			debug($e);
-
 			return response()->json(['message' => 'Something went wrong, try again later'], 500);
 		}
 	}
@@ -37,6 +42,8 @@ class CategoryImageController extends Controller
 	public function destroy(Request $request, $id)
 	{
 		try {
+			$this->authorize('update', Category::class);
+
 			DB::beginTransaction();
 
 			$this->imageService->deleteImage('categories', 'image', $id, $request->imagePath, "");
@@ -45,6 +52,8 @@ class CategoryImageController extends Controller
 
 			DB::commit();
 			return response()->json(['message' => 'Image removed'], 200);
+		} catch (AuthorizationException $e) {
+			return  response()->json(['message' => $e->getMessage()], 403);
 		} catch (\Exception $e) {
 			DB::rollBack();
 			debug($e);
